requirements.txt
fastapi>=0.120,<0.121
sqlalchemy>=2.0,<3
alembic>=1.13,<2
psycopg[binary]>=3.2,<4
httpx>=0.27,<1
starlette>=0.49.1,<0.50
python-jose[cryptography]>=3.3,<4
ecdsa>=0.19.1,<1
uvicorn[standard]>=0.30,<0.31
pydantic>=2.8,<3
redis>=5.0,<6
langgraph>=1.0.8,<2
langchain-core>=1.2.13,<2
langchain-openai>=1.1.9,<2
prometheus-client>=0.20,<1
jsonschema>=4.23,<5
opentelemetry-api>=1.25,<2
opentelemetry-sdk>=1.25,<2
opentelemetry-exporter-otlp>=1.25,<2
opentelemetry-instrumentation-fastapi>=0.46b0,<1
opentelemetry-instrumentation-httpx>=0.46b0,<1
temporalio>=1.7,<2
nats-py>=2.8,<3
